requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29
pydantic>=2.6
pydantic-settings>=2.2
python-multipart>=0.0.9
sqlalchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
celery[redis]>=5.3
redis>=5.0
boto3>=1.34
cryptography>=42.0
PyJWT>=2.8
numpy>=1.26
opencv-python-headless>=4.9
deepface>=0.0.93
tensorflow-cpu>=2.15,<2.20  # 2.20 hasn't shipped binary wheels for some platforms; cap until ecosystem catches up
tf-keras                    # required by DeepFace under tensorflow>=2.16, which dropped bundled keras 2.x
onnxruntime>=1.17
mediapipe>=0.10.9,<0.10.15  # 0.10.15+ removed mediapipe.solutions; pin until we migrate to mediapipe.tasks.vision
slowapi>=0.1.9
structlog>=24.1
httpx>=0.27
